change schema

+++ db/schema.rb
@@ -11,8 +11,9 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema.define(version: 2020_11_20_214026) do
+
   # These are extensions that must be enabled in order to support this database
-  enable_extension "plpgsql" 
+  enable_extension "plpgsql"
 
   create_table "articles", force: :cascade do |t|
     t.string "title"
@@ -28,15 +29,15 @@ ActiveRecord::Schema.define(version: 2020_11_20_214026) do
   create_table "comments", force: :cascade do |t|
     t.string "author_name"
     t.text "body"
-    t.integer "article_id", null: false
+    t.bigint "article_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["article_id"], name: "index_comments_on_article_id"
   end
 
   create_table "taggings", force: :cascade do |t|
-    t.integer "tag_id", null: false
-    t.integer "article_id", null: false
+    t.bigint "tag_id", null: false
+    t.bigint "article_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["article_id"], name: "index_taggings_on_article_id"
